release sept_01_2021

+++ docker/lite/install_dependencies.sh
@@ -11,9 +11,7 @@ FLAVOR="$1"
 export DEBIAN_FRONTEND=noninteractive
 
 KEYSERVERS=(
-    ha.pool.sks-keyservers.net
     keyserver.ubuntu.com
-    hkp://p80.pool.sks-keyservers.net:80
     hkp://keyserver.ubuntu.com:80
 )
 
@@ -49,6 +47,7 @@ BASE_PACKAGES=(
     libatomic1
     libcurl4
     libdbd-mysql-perl
+    libwww-perl
     libev4
     libjemalloc2
     libtcmalloc-minimal4
@@ -92,9 +91,10 @@ mysql57)
     )
     ;;
 mysql80)
-    mysql8_version=8.0.21
+    mysql8_version=8.0.23
     do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/libmysqlclient21_${mysql8_version}-1debian10_amd64.deb /tmp/libmysqlclient21_${mysql8_version}-1debian10_amd64.deb
     do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/mysql-community-client-core_${mysql8_version}-1debian10_amd64.deb /tmp/mysql-community-client-core_${mysql8_version}-1debian10_amd64.deb
+    do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/mysql-community-client-plugins_${mysql8_version}-1debian10_amd64.deb /tmp/mysql-community-client-plugins_${mysql8_version}-1debian10_amd64.deb
     do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/mysql-community-client_${mysql8_version}-1debian10_amd64.deb /tmp/mysql-community-client_${mysql8_version}-1debian10_amd64.deb
     do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/mysql-client_${mysql8_version}-1debian10_amd64.deb /tmp/mysql-client_${mysql8_version}-1debian10_amd64.deb
     do_fetch https://repo.mysql.com/apt/debian/pool/mysql-8.0/m/mysql-community/mysql-community-server-core_${mysql8_version}-1debian10_amd64.deb /tmp/mysql-community-server-core_${mysql8_version}-1debian10_amd64.deb
@@ -103,6 +103,7 @@ mysql80)
     PACKAGES=(
         /tmp/libmysqlclient21_${mysql8_version}-1debian10_amd64.deb
         /tmp/mysql-community-client-core_${mysql8_version}-1debian10_amd64.deb
+        /tmp/mysql-community-client-plugins_${mysql8_version}-1debian10_amd64.deb
         /tmp/mysql-community-client_${mysql8_version}-1debian10_amd64.deb
         /tmp/mysql-client_${mysql8_version}-1debian10_amd64.deb
         /tmp/mysql-community-server-core_${mysql8_version}-1debian10_amd64.deb
@@ -230,6 +231,9 @@ esac
 # Install flavor-specific packages
 apt-get update
 for i in $(seq 1 $MAX_RETRY); do apt-get install -y --no-install-recommends "${PACKAGES[@]}" && break; done
+if [[ "$i" = "$MAX_RETRY" ]]; then
+    exit 1
+fi
 
 # Clean up files we won't need in the final image.
 rm -rf /var/lib/apt/lists/*
