+++ go/test/endtoend/cluster/vttablet_process.go
@@ -18,9 +18,11 @@ limitations under the License.
 package cluster
 
 import (
+	"bufio"
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"os"
@@ -29,6 +31,7 @@ import (
 	"reflect"
 	"strings"
 	"syscall"
+	"testing"
 	"time"
 
 	"vitess.io/vitess/go/mysql"
@@ -68,6 +71,7 @@ type VttabletProcess struct {
 	DbPassword                  string
 	DbPort                      int
 	VreplicationTabletType      string
+	DbFlavor                    string
 	//Extra Args to be set before starting the vttablet process
 	ExtraArgs []string
 
@@ -104,6 +108,9 @@ func (vttablet *VttabletProcess) Setup() (err error) {
 	if *isCoverage {
 		vttablet.proc.Args = append(vttablet.proc.Args, "-test.coverprofile="+getCoveragePath("vttablet.out"))
 	}
+	if *PerfTest {
+		vttablet.proc.Args = append(vttablet.proc.Args, "-pprof", fmt.Sprintf("cpu,waitSig,path=vttablet_pprof_%s", vttablet.Name))
+	}
 
 	if vttablet.SupportsBackup {
 		vttablet.proc.Args = append(vttablet.proc.Args, "-restore_from_backup")
@@ -111,10 +118,13 @@ func (vttablet *VttabletProcess) Setup() (err error) {
 	if vttablet.EnableSemiSync {
 		vttablet.proc.Args = append(vttablet.proc.Args, "-enable_semi_sync")
 	}
+	if vttablet.DbFlavor != "" {
+		vttablet.proc.Args = append(vttablet.proc.Args, fmt.Sprintf("-db_flavor=%s", vttablet.DbFlavor))
+	}
 
 	vttablet.proc.Args = append(vttablet.proc.Args, vttablet.ExtraArgs...)
-
-	errFile, _ := os.Create(path.Join(vttablet.LogDir, vttablet.TabletPath+"-vttablet-stderr.txt"))
+	fname := path.Join(vttablet.LogDir, vttablet.TabletPath+"-vttablet-stderr.txt")
+	errFile, _ := os.Create(fname)
 	vttablet.proc.Stderr = errFile
 
 	vttablet.proc.Env = append(vttablet.proc.Env, os.Environ()...)
@@ -134,7 +144,11 @@ func (vttablet *VttabletProcess) Setup() (err error) {
 	}()
 
 	if vttablet.ServingStatus != "" {
-		if err = vttablet.WaitForTabletType(vttablet.ServingStatus); err != nil {
+		if err = vttablet.WaitForTabletStatus(vttablet.ServingStatus); err != nil {
+			errFileContent, _ := ioutil.ReadFile(fname)
+			if errFileContent != nil {
+				log.Infof("vttablet error:\n%s\n", string(errFileContent))
+			}
 			return fmt.Errorf("process '%s' timed out after 10s (err: %s)", vttablet.Name, err)
 		}
 	}
@@ -198,23 +212,37 @@ func (vttablet *VttabletProcess) GetTabletStatus() string {
 	return ""
 }
 
-// WaitForTabletType waits for 10 second till expected type reached
-func (vttablet *VttabletProcess) WaitForTabletType(expectedType string) error {
-	return vttablet.WaitForTabletTypesForTimeout([]string{expectedType}, 10*time.Second)
+// GetTabletType returns the tablet type as seen in /debug/vars TabletType
+func (vttablet *VttabletProcess) GetTabletType() string {
+	resultMap := vttablet.GetVars()
+	if resultMap != nil {
+		return reflect.ValueOf(resultMap["TabletType"]).String()
+	}
+	return ""
 }
 
-// WaitForTabletTypes waits for 10 second till expected type reached
+// WaitForTabletStatus waits for 10 second till expected status is reached
+func (vttablet *VttabletProcess) WaitForTabletStatus(expectedStatus string) error {
+	return vttablet.WaitForTabletStatusesForTimeout([]string{expectedStatus}, 10*time.Second)
+}
+
+// WaitForTabletStatuses waits for 10 second till one of expected statuses is reached
+func (vttablet *VttabletProcess) WaitForTabletStatuses(expectedStatuses []string) error {
+	return vttablet.WaitForTabletStatusesForTimeout(expectedStatuses, 10*time.Second)
+}
+
+// WaitForTabletTypes waits for 10 second till one of expected statuses is reached
 func (vttablet *VttabletProcess) WaitForTabletTypes(expectedTypes []string) error {
 	return vttablet.WaitForTabletTypesForTimeout(expectedTypes, 10*time.Second)
 }
 
-// WaitForTabletTypesForTimeout waits till the tablet reaches to any of the provided status
-func (vttablet *VttabletProcess) WaitForTabletTypesForTimeout(expectedTypes []string, timeout time.Duration) error {
-	timeToWait := time.Now().Add(timeout)
+// WaitForTabletStatusesForTimeout waits till the tablet reaches to any of the provided statuses
+func (vttablet *VttabletProcess) WaitForTabletStatusesForTimeout(expectedStatuses []string, timeout time.Duration) error {
+	waitUntil := time.Now().Add(timeout)
 	var status string
-	for time.Now().Before(timeToWait) {
+	for time.Now().Before(waitUntil) {
 		status = vttablet.GetTabletStatus()
-		if contains(expectedTypes, status) {
+		if contains(expectedStatuses, status) {
 			return nil
 		}
 		select {
@@ -225,7 +253,27 @@ func (vttablet *VttabletProcess) WaitForTabletTypesForTimeout(expectedTypes []st
 		}
 	}
 	return fmt.Errorf("Vttablet %s, current status = %s, expected status [%s] not reached, details: %v",
-		vttablet.TabletPath, status, strings.Join(expectedTypes, ","), vttablet.GetStatusDetails())
+		vttablet.TabletPath, status, strings.Join(expectedStatuses, ","), vttablet.GetStatusDetails())
+}
+
+// WaitForTabletTypesForTimeout waits till the tablet reaches to any of the provided types
+func (vttablet *VttabletProcess) WaitForTabletTypesForTimeout(expectedTypes []string, timeout time.Duration) error {
+	waitUntil := time.Now().Add(timeout)
+	var tabletType string
+	for time.Now().Before(waitUntil) {
+		tabletType = vttablet.GetTabletType()
+		if contains(expectedTypes, tabletType) {
+			return nil
+		}
+		select {
+		case err := <-vttablet.exit:
+			return fmt.Errorf("process '%s' exited prematurely (err: %s)", vttablet.Name, err)
+		default:
+			time.Sleep(300 * time.Millisecond)
+		}
+	}
+	return fmt.Errorf("Vttablet %s, current type = %s, expected type [%s] not reached, status details: %v",
+		vttablet.TabletPath, tabletType, strings.Join(expectedTypes, ","), vttablet.GetStatusDetails())
 }
 
 func contains(arr []string, str string) bool {
@@ -281,8 +329,14 @@ func (vttablet *VttabletProcess) getVarValue(keyname string) string {
 	return fmt.Sprintf("%v", object)
 }
 
-// TearDown shuts down the running vttablet service
+// TearDown shuts down the running vttablet service and fails after 10 seconds
 func (vttablet *VttabletProcess) TearDown() error {
+	return vttablet.TearDownWithTimeout(10 * time.Second)
+}
+
+// TearDownWithTimeout shuts down the running vttablet service and fails once the given
+// duration has elapsed.
+func (vttablet *VttabletProcess) TearDownWithTimeout(timeout time.Duration) error {
 	if vttablet.proc == nil || vttablet.exit == nil {
 		return nil
 	}
@@ -294,9 +348,12 @@ func (vttablet *VttabletProcess) TearDown() error {
 		vttablet.proc = nil
 		return nil
 
-	case <-time.After(10 * time.Second):
-		vttablet.proc.Process.Kill()
-		vttablet.proc = nil
+	case <-time.After(timeout):
+		proc := vttablet.proc
+		if proc != nil {
+			vttablet.proc.Process.Kill()
+			vttablet.proc = nil
+		}
 		return <-vttablet.exit
 	}
 }
@@ -314,11 +371,15 @@ func (vttablet *VttabletProcess) QueryTablet(query string, keyspace string, useD
 		keyspace = ""
 	}
 	dbParams := NewConnParams(vttablet.DbPort, vttablet.DbPassword, path.Join(vttablet.Directory, "mysql.sock"), keyspace)
-	return executeQuery(dbParams, query)
+	conn, err := vttablet.conn(&dbParams)
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+	return executeQuery(conn, query)
 }
 
-// QueryTabletWithDB lets you execute query on a specific DB in this tablet and get the result
-func (vttablet *VttabletProcess) QueryTabletWithDB(query string, dbname string) (*sqltypes.Result, error) {
+func (vttablet *VttabletProcess) defaultConn(dbname string) (*mysql.Conn, error) {
 	dbParams := mysql.ConnParams{
 		Uname:      "vt_dba",
 		UnixSocket: path.Join(vttablet.Directory, "mysql.sock"),
@@ -327,18 +388,26 @@ func (vttablet *VttabletProcess) QueryTabletWithDB(query string, dbname string) 
 	if vttablet.DbPassword != "" {
 		dbParams.Pass = vttablet.DbPassword
 	}
-	return executeQuery(dbParams, query)
+	return vttablet.conn(&dbParams)
 }
 
-func executeQuery(dbParams mysql.ConnParams, query string) (*sqltypes.Result, error) {
+func (vttablet *VttabletProcess) conn(dbParams *mysql.ConnParams) (*mysql.Conn, error) {
 	ctx := context.Background()
-	dbConn, err := mysql.Connect(ctx, &dbParams)
+	return mysql.Connect(ctx, dbParams)
+}
+
+// QueryTabletWithDB lets you execute query on a specific DB in this tablet and get the result
+func (vttablet *VttabletProcess) QueryTabletWithDB(query string, dbname string) (*sqltypes.Result, error) {
+	conn, err := vttablet.defaultConn(dbname)
 	if err != nil {
 		return nil, err
 	}
-	defer dbConn.Close()
-	qr, err := dbConn.ExecuteFetch(query, 10000, true)
-	return qr, err
+	defer conn.Close()
+	return executeQuery(conn, query)
+}
+
+func executeQuery(dbConn *mysql.Conn, query string) (*sqltypes.Result, error) {
+	return dbConn.ExecuteFetch(query, 10000, true)
 }
 
 // GetDBVar returns first matching database variable's value
@@ -360,6 +429,96 @@ func (vttablet *VttabletProcess) getDBSystemValues(placeholder string, value str
 		return fmt.Sprintf("%s", output.Rows[0][1].ToBytes()), nil
 	}
 	return "", nil
+}
+
+// ToggleProfiling enables or disables the configured CPU profiler on this vttablet
+func (vttablet *VttabletProcess) ToggleProfiling() error {
+	return vttablet.proc.Process.Signal(syscall.SIGUSR1)
+}
+
+// WaitForVReplicationToCatchup waits for "workflow" to finish copying
+func (vttablet *VttabletProcess) WaitForVReplicationToCatchup(t testing.TB, workflow, database string, duration time.Duration) {
+	queries := [3]string{
+		fmt.Sprintf(`select count(*) from _vt.vreplication where workflow = "%s" and db_name = "%s" and pos = ''`, workflow, database),
+		"select count(*) from information_schema.tables where table_schema='_vt' and table_name='copy_state' limit 1;",
+		fmt.Sprintf(`select count(*) from _vt.copy_state where vrepl_id in (select id from _vt.vreplication where workflow = "%s" and db_name = "%s" )`, workflow, database),
+	}
+	results := [3]string{"[INT64(0)]", "[INT64(1)]", "[INT64(0)]"}
+
+	conn, err := vttablet.defaultConn("")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Close()
+
+	var lastChecked time.Time
+	for ind, query := range queries {
+		waitDuration := 500 * time.Millisecond
+		for duration > 0 {
+			log.Infof("Executing query %s on %s", query, vttablet.Name)
+			lastChecked = time.Now()
+			qr, err := conn.ExecuteFetch(query, 1000, true)
+			if err != nil {
+				t.Fatal(err)
+			}
+			if qr != nil && qr.Rows != nil && len(qr.Rows) > 0 && fmt.Sprintf("%v", qr.Rows[0]) == string(results[ind]) {
+				break
+			} else {
+				log.Infof("In WaitForVReplicationToCatchup: %s %+v", query, qr.Rows)
+			}
+			time.Sleep(waitDuration)
+			duration -= waitDuration
+		}
+		if duration <= 0 {
+			t.Fatalf("WaitForVReplicationToCatchup timed out for workflow %s, keyspace %s", workflow, database)
+		}
+	}
+	log.Infof("WaitForVReplicationToCatchup succeeded at %v", lastChecked)
+}
+
+// BulkLoad performs a bulk load of rows into a given vttablet.
+func (vttablet *VttabletProcess) BulkLoad(t testing.TB, db, table string, bulkInsert func(io.Writer)) {
+	tmpbulk, err := ioutil.TempFile(path.Join(vttablet.Directory, "tmp"), "bulk_load")
+	if err != nil {
+		t.Fatalf("failed to create tmp file for loading: %v", err)
+	}
+	defer os.Remove(tmpbulk.Name())
+
+	log.Infof("create temporary file for bulk loading %q", tmpbulk.Name())
+	bufStart := time.Now()
+
+	bulkBuffer := bufio.NewWriter(tmpbulk)
+	bulkInsert(bulkBuffer)
+	bulkBuffer.Flush()
+
+	pos, _ := tmpbulk.Seek(0, 1)
+	bufFinish := time.Now()
+	log.Infof("bulk loading %d bytes from %q...", pos, tmpbulk.Name())
+
+	if err := tmpbulk.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	conn, err := vttablet.defaultConn("vt_" + db)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Close()
+
+	query := fmt.Sprintf("LOAD DATA INFILE '%s' INTO TABLE `%s` FIELDS TERMINATED BY ',' ENCLOSED BY '\"'", tmpbulk.Name(), table)
+	_, err = conn.ExecuteFetch(query, 1, false)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	end := time.Now()
+	log.Infof("bulk insert successful (write tmp file = %v, mysql bulk load = %v, total = %v",
+		bufFinish.Sub(bufStart), end.Sub(bufFinish), end.Sub(bufStart))
+}
+
+// IsShutdown returns whether a vttablet is shutdown or not
+func (vttablet *VttabletProcess) IsShutdown() bool {
+	return vttablet.proc == nil
 }
 
 // VttabletProcessInstance returns a VttabletProcess handle for vttablet process
